+++ web-console/src/lib/components/connectors/dialogs/KafkaOutputConnector.tsx
@@ -1,19 +1,15 @@
 // A create/update dialog for a Kafka output connector.
 'use client'
 
-import TabKafkaNameAndDesc from '$lib/components/connectors/dialogs/tabs/kafka/TabKafkaNameAndDesc'
-import TabFooter from '$lib/components/connectors/dialogs/tabs/TabFooter'
-import TabLabel from '$lib/components/connectors/dialogs/tabs/TabLabel'
+import { TabKafkaNameAndDesc } from '$lib/components/connectors/dialogs/tabs/kafka/TabKafkaNameAndDesc'
+import { TabFooter } from '$lib/components/connectors/dialogs/tabs/TabFooter'
+import { TabLabel } from '$lib/components/connectors/dialogs/tabs/TabLabel'
 import { connectorTransportName, parseKafkaOutputSchema, parseKafkaOutputSchemaConfig } from '$lib/functions/connectors'
-import {
-  authFields,
-  authParamsSchema,
-  defaultUiAuthParams,
-  prepareAuthData
-} from '$lib/functions/kafka/authParamsSchema'
+import { authFields, authParamsSchema, defaultLibrdkafkaAuthOptions } from '$lib/functions/kafka/authParamsSchema'
+import { LibrdkafkaOptionType, toKafkaConfig } from '$lib/functions/kafka/librdkafkaOptions'
 import { useConnectorRequest } from '$lib/services/connectors/dialogs/SubmitHandler'
 import { ConnectorType, Direction } from '$lib/types/connectors'
-import ConnectorDialogProps from '$lib/types/connectors/ConnectorDialogProps'
+import { ConnectorDialogProps } from '$lib/types/connectors/ConnectorDialogProps'
 import { useEffect, useState } from 'react'
 import { FieldErrors } from 'react-hook-form'
 import { FormContainer } from 'react-hook-form-mui'
@@ -38,31 +34,36 @@ import Typography from '@mui/material/Typography'
 
 import { GenericEditorForm } from './tabs/GenericConnectorForm'
 import { TabKafkaAuth } from './tabs/kafka/TabKafkaAuth'
-import TabkafkaOutputDetails from './tabs/kafka/TabKafkaOutputDetails'
+import { TabKafkaOutputDetails } from './tabs/kafka/TabKafkaOutputDetails'
 import TabOutputFormatDetails from './tabs/TabOutputFormatDetails'
 import Transition from './tabs/Transition'
 
 const schema = va.object({
-  name: va.nonOptional(va.string()),
+  name: va.nonOptional(va.string([va.minLength(1, 'Specify connector name')])),
   description: va.optional(va.string(), ''),
-  config: va.intersect([
+  transport: va.intersect([
     va.object({
-      bootstrap_servers: va.nonOptional(va.string()),
-      topic: va.optional(va.string(), ''),
-      format_name: va.nonOptional(va.enumType(['json', 'csv'])),
-      json_array: va.nonOptional(va.boolean()),
-      transport_user_config: va.optional(va.any()),
-      format_user_config: va.optional(va.any())
+      bootstrap_servers: va.nonOptional(
+        va.array(va.string([va.minLength(1, 'Specify at least one server')]), [
+          va.minLength(1, 'Specify at least one server')
+        ])
+      ),
+      topic: va.optional(va.string(), '')
     }),
     authParamsSchema
-  ])
+  ]),
+  format: va.object({
+    format_name: va.nonOptional(va.picklist(['json', 'csv'])),
+    json_array: va.nonOptional(va.boolean())
+  })
 })
 
 export type KafkaOutputSchema = va.Input<typeof schema>
 
 export const KafkaOutputConnectorDialog = (props: ConnectorDialogProps) => {
+  const tabs = ['detailsTab', 'sourceTab', 'authTab', 'formatTab'] as const
   const [rawJSON, setRawJSON] = useState(false)
-  const [activeTab, setActiveTab] = useState<string>('detailsTab')
+  const [activeTab, setActiveTab] = useState<(typeof tabs)[number]>('detailsTab')
   const [curValues, setCurValues] = useState<KafkaOutputSchema | undefined>(undefined)
 
   // Initialize the form either with values from the passed in connector
@@ -75,17 +76,19 @@ export const KafkaOutputConnectorDialog = (props: ConnectorDialogProps) => {
   const defaultValues: KafkaOutputSchema = {
     name: '',
     description: '',
-    config: {
-      bootstrap_servers: '',
+    transport: {
+      bootstrap_servers: [''],
       topic: '',
+      ...defaultLibrdkafkaAuthOptions
+    },
+    format: {
       format_name: 'json',
-      json_array: false,
-      ...defaultUiAuthParams
+      json_array: false
     }
   }
 
   const handleClose = () => {
-    setActiveTab(tabList[0])
+    setActiveTab(tabs[0])
     props.setShow(false)
   }
 
@@ -93,26 +96,23 @@ export const KafkaOutputConnectorDialog = (props: ConnectorDialogProps) => {
   const prepareData = (data: KafkaOutputSchema) => ({
     name: data.name,
     description: data.description,
-    config: normalizeConfig(data.config)
+    config: normalizeConfig(data)
   })
 
-  const normalizeConfig = (data: KafkaOutputSchema['config']) => ({
+  const normalizeConfig = (data: {
+    transport: Record<string, LibrdkafkaOptionType>
+    format: Record<string, string | boolean>
+  }) => ({
     transport: {
       name: connectorTransportName(ConnectorType.KAFKA_OUT),
-      config: {
-        ...data.transport_user_config,
-        'bootstrap.servers': data.bootstrap_servers,
-        topic: data.topic,
-        ...prepareAuthData(data)
-      }
+      config: toKafkaConfig(data.transport)
     },
     format: {
-      name: data.format_name,
+      name: data.format.format_name,
       config: {
-        ...data.format_user_config,
-        ...(data.format_name === 'json'
+        ...(data.format.format_name === 'json'
           ? {
-              array: data.json_array
+              array: data.format.json_array
             }
           : {})
       }
@@ -122,24 +122,23 @@ export const KafkaOutputConnectorDialog = (props: ConnectorDialogProps) => {
   const onSubmit = useConnectorRequest(props.connector, prepareData, props.onSuccess, handleClose)
 
   // If there is an error, switch to the earliest tab with an error
-  const handleErrors = ({ name, description, config }: FieldErrors<KafkaOutputSchema>) => {
+  const handleErrors = ({ name, description, transport, format }: FieldErrors<KafkaOutputSchema>) => {
     if (!props.show) {
       return
     }
     if (name || description) {
       setActiveTab('detailsTab')
-    } else if (config?.bootstrap_servers || config?.topic) {
+    } else if (transport?.bootstrap_servers || transport?.topic) {
       setActiveTab('sourceTab')
-    } else if (config && authFields.some(f => f in config)) {
+    } else if (transport && authFields.some(f => f in transport)) {
       setActiveTab('authTab')
-    } else if (config?.format_name || config?.json_array) {
+    } else if (format?.format_name || format?.json_array) {
       setActiveTab('formatTab')
     }
   }
 
-  const tabList = ['detailsTab', 'sourceTab', 'authTab', 'formatTab']
   const tabFooter = (
-    <TabFooter submitButton={props.submitButton} activeTab={activeTab} setActiveTab={setActiveTab} tabsArr={tabList} />
+    <TabFooter submitButton={props.submitButton} activeTab={activeTab} setActiveTab={setActiveTab} tabs={tabs} />
   )
 
   const jsonSwitch = (
@@ -192,8 +191,8 @@ export const KafkaOutputConnectorDialog = (props: ConnectorDialogProps) => {
               <GenericEditorForm
                 disabled={props.disabled}
                 direction={Direction.OUTPUT}
-                configFromText={t => parseKafkaOutputSchemaConfig(JSON.parse(t))}
-                configToText={c => JSON.stringify(normalizeConfig(c as any), undefined, '\t')}
+                configFromText={text => parseKafkaOutputSchemaConfig(JSON.parse(text))}
+                configToText={config => JSON.stringify(normalizeConfig(config), undefined, '\t')}
               />
               <Box sx={{ display: 'flex', justifyContent: 'end' }}>{props.submitButton}</Box>
             </Box>
@@ -204,7 +203,7 @@ export const KafkaOutputConnectorDialog = (props: ConnectorDialogProps) => {
                   {jsonSwitch}
                   <TabList
                     orientation='vertical'
-                    onChange={(e, newValue: string) => setActiveTab(newValue)}
+                    onChange={(e, newValue: (typeof tabs)[number]) => setActiveTab(newValue)}
                     sx={{
                       border: 0,
                       minWidth: 200,
@@ -283,14 +282,14 @@ export const KafkaOutputConnectorDialog = (props: ConnectorDialogProps) => {
                   value='sourceTab'
                   sx={{ border: 0, boxShadow: 0, width: '100%', backgroundColor: 'transparent' }}
                 >
-                  <TabkafkaOutputDetails disabled={props.disabled} />
+                  <TabKafkaOutputDetails disabled={props.disabled} parentName='transport' />
                   {tabFooter}
                 </TabPanel>
                 <TabPanel
                   value='authTab'
                   sx={{ border: 0, boxShadow: 0, width: '100%', backgroundColor: 'transparent' }}
                 >
-                  <TabKafkaAuth disabled={props.disabled} />
+                  <TabKafkaAuth disabled={props.disabled} parentName={'transport'} />
                   {tabFooter}
                 </TabPanel>
                 <TabPanel