+++ web-console/src/lib/components/connectors/dialogs/elements/kafka/Egress.tsx
@@ -1,0 +1,16 @@
+import { TextFieldElement } from 'react-hook-form-mui'
+
+export const KafkaTopicElement = (props: { disabled?: boolean; parentName: string }) => (
+  <TextFieldElement
+    name={props.parentName + '.topic'}
+    label='Topic Name'
+    size='small'
+    fullWidth
+    placeholder='my-topic'
+    aria-describedby='validation-topic'
+    disabled={props.disabled}
+    inputProps={{
+      'data-testid': 'input-topic'
+    }}
+  />
+)